Change a number of USB arguments/fields/variables to size_t/ssize_t (from int). (dm)
git-svn-id: svn://mielke.cc/main/brltty@5253 91a5dbb7-01b9-0310-9b5f-b28072856b6e

+++ Programs/usb_bsd.h
@@ -164,7 +164,7 @@ usbSubmitRequest (
   UsbDevice *device,
   unsigned char endpointAddress,
   void *buffer,
-  int length,
+  size_t length,
   void *context
 ) {
   errno = ENOSYS;
@@ -199,7 +199,7 @@ usbReadEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   int count = -1;
@@ -225,7 +225,7 @@ usbWriteEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   const void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   UsbEndpoint *endpoint = usbGetOutputEndpoint(device, endpointNumber);
