+++ Programs/usb_none.c
@@ -115,7 +115,7 @@ usbSubmitRequest (
   UsbDevice *device,
   unsigned char endpointAddress,
   void *buffer,
-  int length,
+  size_t length,
   void *context
 ) {
   errno = ENOSYS;
@@ -150,7 +150,7 @@ usbReadEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   errno = ENOSYS;
@@ -163,7 +163,7 @@ usbWriteEndpoint (
   UsbDevice *device,
   unsigned char endpointNumber,
   const void *buffer,
-  unsigned int length,
+  size_t length,
   int timeout
 ) {
   errno = ENOSYS;
